admin_portal/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
firebase-admin==6.9.0
PyJWT==2.10.1
python-multipart
python-dotenv==1.0.1
email-validator==2.2.0